src/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.30.0
pytz==2023.3
email-validator==2.0.0

